create a table
create a table for 'book a table'
