+++ CafeApp/CafeApp/App.config
@@ -6,5 +6,8 @@
         <add name="CafeApp.Properties.Settings.RestaurantsConnectionString"
             connectionString="Data Source=(LocalDB)\MSSQLLocalDB;AttachDbFilename=|DataDirectory|\Restaurants.mdf;Integrated Security=True"
             providerName="System.Data.SqlClient" />
+        <add name="CafeApp.Properties.Settings.BookDatabaseConnectionString"
+            connectionString="Data Source=(LocalDB)\MSSQLLocalDB;AttachDbFilename=|DataDirectory|\BookDatabase.mdf;Integrated Security=True"
+            providerName="System.Data.SqlClient" />
     </connectionStrings>
 </configuration>
